+++ packages/stepper/src/lib/StepLabel.tsx
@@ -1,12 +1,22 @@
 import MuiStepLabel, { StepLabelProps as MuiStepLabelProps } from '@mui/material/StepLabel';
 
 import { StepIcon, StepIconProps } from './StepIcon';
+import { Typography } from '@availity/mui-typography';
 
 export type StepLabelProps = Pick<StepIconProps, 'warning'> & MuiStepLabelProps;
 
-const StepLabel = ({ children, error, warning, ...rest }: StepLabelProps): JSX.Element => {
+const StepLabel = ({ children, error, optional, warning, ...rest }: StepLabelProps): JSX.Element => {
+  if (typeof optional === 'string') {
+    optional = <Typography variant="caption" children={optional} />;
+  }
   return (
-    <MuiStepLabel StepIconProps={{ error, warning }} StepIconComponent={StepIcon} error={error} {...rest}>
+    <MuiStepLabel
+      StepIconProps={{ error, warning }}
+      StepIconComponent={StepIcon}
+      error={error}
+      optional={optional}
+      {...rest}
+    >
       {children}
     </MuiStepLabel>
   );