+++ packages/theme/src/lib/light-theme.ts
@@ -74,8 +74,11 @@ export const lightTheme = {
       inputFocus: tokens.borderInputFocus,
       inverse: tokens.borderInverse,
       inverseSecondary: tokens.borderInverseSecondary,
+      hover: tokens.colorInfoMain20,
     },
     background: {
+      primary: tokens.colorPrimaryMain20,
+      secondary: '#393B4621',
       canvas: tokens.colorBackgroundCanvas,
       paper: tokens.colorBackgroundPaper,
       header: tokens.colorBackgroundHeader,
@@ -434,7 +437,7 @@ export const lightTheme = {
           '.MuiListItemButton-root': {
             borderRadius: '0 4px 4px 0',
             '&:active': {
-              backgroundColor: tokens.colorActionFocus
+              backgroundColor: tokens.colorActionFocus,
             },
             '&.Mui-focusVisible': {
               border: `2px solid ${tokens.colorPrimaryMain}`,
@@ -540,11 +543,12 @@ export const lightTheme = {
               '.MuiAccordionSummary-root:not(.Mui-focusVisible, :hover, :active)': {
                 backgroundColor: 'inherit',
               },
-              '> .MuiCollapse-root > .MuiCollapse-wrapper > .MuiCollapse-wrapperInner > .MuiAccordion-region > .MuiAccordionDetails-root:first-of-type': {
-                borderTop: 'none'
-              }
+              '> .MuiCollapse-root > .MuiCollapse-wrapper > .MuiCollapse-wrapperInner > .MuiAccordion-region > .MuiAccordionDetails-root:first-of-type':
+                {
+                  borderTop: 'none',
+                },
             },
-          }
+          },
         },
       },
     },
@@ -552,18 +556,18 @@ export const lightTheme = {
       styleOverrides: {
         root: {
           padding: '1.25rem',
-          flexWrap: 'wrap'
+          flexWrap: 'wrap',
         },
         spacing: {
-          rowGap: tokens.spacingBase
-        }
+          rowGap: tokens.spacingBase,
+        },
       },
     },
     MuiAccordionDetails: {
       styleOverrides: {
         root: {
           minHeight: '1px',
-          padding: '1.25rem'
+          padding: '1.25rem',
         },
       },
     },
@@ -595,7 +599,7 @@ export const lightTheme = {
           order: 2,
           fontWeight: tokens.fontWeightsBold,
           margin: '12px 0px',
-          gap: 1
+          gap: 1,
         },
         expandIconWrapper: {
           order: 1,
@@ -624,11 +628,11 @@ export const lightTheme = {
             margin: '0px -4px',
           },
           '.MuiIconButton-root': {
-            marginRight: '-8px'
+            marginRight: '-8px',
           },
           margin: '-2px 0px -2px auto',
           padding: '0 0 0 16px',
-          gap: '16px'
+          gap: '16px',
         },
         icon: ({ ownerState }: { ownerState: any }) => ({
           ...(ownerState.severity === 'success' && {
@@ -713,14 +717,14 @@ export const lightTheme = {
             height: '2.25rem',
             '.MuiSelect-avExpandIcon': {
               transform: 'unset',
-              top: 'auto'
+              top: 'auto',
             },
           },
         },
         popupIndicator: {
           color: tokens.borderInput,
 
-          padding: '.75rem'
+          padding: '.75rem',
         },
         popupIndicatorOpen: {
           transform: 'scaleY(-1)',
@@ -977,7 +981,7 @@ export const lightTheme = {
         sizeSmall: {
           verticalAlign: 'text-bottom',
           borderRadius: '50px',
-          height: 'min-content'
+          height: 'min-content',
         },
         sizeMedium: {
           fontWeight: tokens.fontWeightsRegular,
@@ -995,9 +999,9 @@ export const lightTheme = {
     MuiDateCalendar: {
       styleOverrides: {
         root: {
-          width: '310px'
-        }
-      }
+          width: '310px',
+        },
+      },
     },
     MuiDatePicker: {
       styleOverrides: {
@@ -1021,51 +1025,51 @@ export const lightTheme = {
         PaperProps: {
           variant: 'elevation',
           elevation: 4,
-        }
+        },
       },
       styleOverrides: {
         maxWidthSm: {
-          maxWidth: '25rem'
+          maxWidth: '25rem',
         },
         maxWidthMd: {
-          maxWidth: '39.75rem'
+          maxWidth: '39.75rem',
         },
         maxWidthLg: {
-          maxWidth: '58.75rem'
+          maxWidth: '58.75rem',
         },
         maxWidthXl: {
-          maxWidth: '71.25rem'
+          maxWidth: '71.25rem',
         },
         avCloseButton: {
-          "~ .MuiDialogTitle-root": {
-            marginRight: '2.5rem'
-          }
-        }
-      }
+          '~ .MuiDialogTitle-root': {
+            marginRight: '2.5rem',
+          },
+        },
+      },
     },
     MuiDialogActions: {
       styleOverrides: {
         root: {
           padding: '1rem',
-          flexWrap: 'wrap'
+          flexWrap: 'wrap',
         },
         spacing: {
           marginBottom: '-8px',
           '.MuiButton-root': {
-            marginBottom: '8px'
-          }
-        }
-      }
+            marginBottom: '8px',
+          },
+        },
+      },
     },
     MuiDialogContent: {
       defaultProps: {
-        dividers: true
+        dividers: true,
       },
       styleOverrides: {
         root: {
-          padding: '1rem'
-        }
-      }
+          padding: '1rem',
+        },
+      },
     },
     MuiDialogTitle: {
       styleOverrides: {
@@ -1074,9 +1078,9 @@ export const lightTheme = {
           paddingLeft: '1rem',
           minHeight: '3.5rem',
           display: 'flex',
-          alignItems: 'center'
-        }
-      }
+          alignItems: 'center',
+        },
+      },
     },
     MuiFormControl: {
       defaultProps: {
@@ -1390,14 +1394,14 @@ export const lightTheme = {
             top: '1.75rem',
           },
           '&.MuiListItem-padding:not(.MuiListItem-dense), &:not(.MuiListItem-padding):not(.MuiListItem-dense) .MuiListItemButton-root:not(.MuiListItemButton-dense)':
-            {
-              paddingTop: '.75rem',
-              paddingBottom: '.75rem',
-              '> .MuiIconButton-root, > .MuiChip-root': {
-                marginTop: '1rem',
-                marginBottom: '-1rem',
-              },
+          {
+            paddingTop: '.75rem',
+            paddingBottom: '.75rem',
+            '> .MuiIconButton-root, > .MuiChip-root': {
+              marginTop: '1rem',
+              marginBottom: '-1rem',
             },
+          },
           '&.MuiListItem-dense': {
             '.MuiListItemButton-root > .MuiChip-root': {
               marginTop: '1rem',
@@ -1484,9 +1488,9 @@ export const lightTheme = {
             backgroundColor: tokens.colorPrimaryMain,
             color: tokens.colorTextInverse,
             '.MuiListItemText-root, .MuiListItemText-root *, .MuiListItemSubheader-root, .MuiListItemSubheader-root *, .MuiListItemIcon-root, .MuiListItemIcon-root *, .MuiListItemSecondaryAction-root, .MuiIconButton-root':
-              {
-                color: 'inherit',
-              },
+            {
+              color: 'inherit',
+            },
             '&.Mui-focusVisible': {
               backgroundColor: tokens.colorPrimaryDark,
               boxShadow: `inset 0 0 0px 1px ${tokens.colorPrimaryDark}, inset 0 0 0px 2px ${tokens.colorCommonWhite}`,
@@ -1717,17 +1721,19 @@ export const lightTheme = {
     },
     MuiSnackbar: {
       styleOverrides: {
-        root: ({ theme } : { theme: Theme} ) => ({
+        root: ({ theme }: { theme: Theme }) => ({
           [theme.breakpoints.down('sm')]: {
-              // full width
-              left: '24px',
-              right: '24px',
-            '&.MuiSnackbar-anchorOriginBottomCenter, &.MuiSnackbar-anchorOriginBottomLeft, &.MuiSnackbar-anchorOriginBottomRight': {
-              bottom: '24px',
-            },
-            '&.MuiSnackbar-anchorOriginTopCenter, &.MuiSnackbar-anchorOriginTopLeft, &.MuiSnackbar-anchorOriginTopRight': {
-              top: '24px',
-            },
+            // full width
+            left: '24px',
+            right: '24px',
+            '&.MuiSnackbar-anchorOriginBottomCenter, &.MuiSnackbar-anchorOriginBottomLeft, &.MuiSnackbar-anchorOriginBottomRight':
+              {
+                bottom: '24px',
+              },
+            '&.MuiSnackbar-anchorOriginTopCenter, &.MuiSnackbar-anchorOriginTopLeft, &.MuiSnackbar-anchorOriginTopRight':
+              {
+                top: '24px',
+              },
           },
           [theme.breakpoints.up('sm')]: {
             maxWidth: '528px',
@@ -1738,7 +1744,7 @@ export const lightTheme = {
           },
           '.MuiAlert-root, .MuiSnackbarContent-root': {
             boxShadow: tokens.shadows24,
-          }
+          },
         }),
         anchorOriginBottomCenter: {
           bottom: '40px',
@@ -1769,39 +1775,104 @@ export const lightTheme = {
         root: {
           // info alert styling
           backgroundColor: tokens.colorBlue50,
-          color: tokens.colorTextPrimary
-        }
+          color: tokens.colorTextPrimary,
+        },
+      },
+    },
+    MuiStep: {
+      styleOverrides: {
+        root: {
+          minWidth: '71px',
+          whiteSpace: 'wrap',
+          '&:not(:first-of-type)': {
+            '.MuiStepLabel-vertical': {
+              '&::after': {
+                content: '""',
+                borderLeft: `4px solid ${tokens.colorGrey200}`,
+                position: 'absolute',
+                left: '26px',
+                bottom: 'calc(50% + 14px)',
+                top: 0,
+              },
+            }
+          },
+          '&:not(:last-of-type)': {
+            '.MuiStepLabel-vertical': {
+              ' &::before': {
+                content: '""',
+                borderLeft: `4px solid ${tokens.colorGrey200}`,
+                position: 'absolute',
+                left: '26px',
+                bottom: 0,
+                top: 'calc(50% + 14px)'
+              }
+            }
+          },
+          '&.Mui-completed': {
+            '.MuiStepLabel-vertical': {
+              '&::before, &::after': {
+                borderColor: tokens.colorSuccessMain,
+              }
+            },
+          },
+          '&.Mui-active': {
+            '.MuiStepLabel-vertical': {
+              '&::before, &::after': {
+                borderColor: tokens.colorSuccessMain,
+              }
+            },
+          }
+        },
+      }
+    },
+    MuiStepButton: {
+      styleOverrides: {
+        root: {
+          borderRadius: 8,
+          margin: 0,
+          padding: 0,
+          '&:hover': {
+            backgroundColor: tokens.colorActionHover,
+          },
+          '&:active': {
+            backgroundColor: "#393B4626"
+          },
+          '&.Mui-focusVisible': {
+            outline: `2px solid ${tokens.borderInputFocus}`
+          },
+        },
       }
     },
     MuiStepIcon: {
       styleOverrides: {
         root: {
           color: tokens.colorCommonWhite,
-          border: `solid ${tokens.colorSuccessMain}`,
+          outline: `2px solid ${tokens.colorSuccessMain}`,
+          outlineOffset: '-2px',
           borderRadius: '50%',
           height: '24px',
           width: '24px',
           '&.Mui-active': {
             color: tokens.colorSuccessMain,
-            border: 'none',
+            outline: 'none',
             '.MuiStepIcon-text': {
               fill: tokens.colorCommonWhite,
             },
           },
           '&.Mui-completed': {
             color: tokens.colorSuccessMain,
-            border: 'none',
+            outline: 'none',
           },
           '&.Mui-error': {
-            border: 'none',
+            outline: 'none',
           },
           '&.MuiSvgIcon-colorWarning': {
-            border: 'none',
+            outline: 'none',
           },
         },
         text: {
           fill: tokens.colorTextSuccess,
-          fontSize: '18px',
+          fontSize: '14px',
         },
       },
     },
@@ -1811,19 +1882,26 @@ export const lightTheme = {
           display: 'flex',
           alignSelf: 'center',
           justifyContent: 'center',
+          padding: '24px 0',
         },
         iconContainer: {
           '.MuiSvgIcon-colorWarning': {
             color: tokens.colorWarningDark,
           },
         },
+        vertical: {
+          padding: '20px 16px',
+          position: 'relative',
+        }
       },
     },
     MuiStepConnector: {
       styleOverrides: {
         root: {
-          top: '8px',
+          top: '34px',
           color: tokens.colorGrey200,
+          left: 'calc(-50% + 14px)',
+          right: 'calc(50% + 14px)',
           '&.Mui-completed': {
             color: tokens.colorSuccessMain,
           },
@@ -1835,7 +1913,30 @@ export const lightTheme = {
           borderColor: 'inherit',
           borderWidth: '4px',
         },
+        vertical: {
+          display: 'none',
+          '&.Mui-active + .MuiStep-vertical .MuiStepLabel-vertical::after': {
+            borderColor: tokens.colorSuccessMain,
+          }
+        },
       },
+    },
+    MuiStepper: {
+      styleOverrides: {
+        root: {
+          flexGrow: 1,
+          overflowX: 'auto',
+          overflowY: 'hidden',
+          whiteSpace: 'nowrap',
+          scrollbarWidth: 'none',
+          paddingTop: '2px',
+          paddingBottom: '2px'
+        },
+        vertical: {
+          alignItems: 'flex-start',
+          overflow: 'visible'
+        }
+      }
     },
     MuiSvgIcon: {
       defaultProps: {
@@ -1849,17 +1950,23 @@ export const lightTheme = {
           color: tokens.colorWarningDark,
         },
         fontSizeXxsmall: {
-          fontSize: '.75rem',
+          fontSize: tokens.iconSizeXxsmall,
         },
         fontSizeXsmall: {
-          fontSize: '1rem',
+          fontSize: tokens.iconSizeXsmall,
+        },
+        fontSizeSmall: {
+          fontSize: tokens.iconSizeSmall,
+        },
+        fontSizeMedium: {
+          fontSize: tokens.iconSizeMedium,
         },
         fontSizeLarge: {
-          fontSize: '2rem',
+          fontSize: tokens.iconSizeLarge,
         },
         fontSizeXlarge: {
-          fontSize: '2.25rem',
-        },
+          fontSize: tokens.iconSizeXlarge,
+        }
       },
     },
     MuiSwitch: {
@@ -1892,11 +1999,11 @@ export const lightTheme = {
             '&.Mui-checked': {
               ' + .MuiSwitch-track': {
                 backgroundColor: tokens.colorPrimaryLight,
-                opacity: 1
+                opacity: 1,
               },
               '&::hover': {
                 '+ .MuiSwitch-track': {
-                  backgroundColor: tokens.colorPrimaryDark
+                  backgroundColor: tokens.colorPrimaryDark,
                 },
               },
             },
@@ -1904,21 +2011,21 @@ export const lightTheme = {
               color: tokens.colorCommonWhite,
               '+ .MuiSwitch-track': {
                 backgroundColor: tokens.colorTextDisabled,
-                opacity: 1
-              }
-            }
+                opacity: 1,
+              },
+            },
           },
         },
         thumb: {
           height: '16px',
           width: '16px',
           boxShadow: 'unset',
-          backgroundColor: tokens.colorCommonWhite
+          backgroundColor: tokens.colorCommonWhite,
         },
         track: {
           borderRadius: '10px',
           background: tokens.colorSecondaryLight,
-          opacity: 'unset'
+          opacity: 'unset',
         },
         sizeSmall: {
           height: '16px',
@@ -1926,30 +2033,30 @@ export const lightTheme = {
           '.MuiSwitch-switchBase': {
             '&.Mui-checked': {
               transform: 'translateX(16px)',
-            }
+            },
           },
           '.MuiSwitch-thumb': {
             height: '12px',
-            width: '12px'
+            width: '12px',
           },
           '.MuiSwitch-track': {
-            borderRadius: '8px'
-          }
+            borderRadius: '8px',
+          },
         },
-      }
+      },
     },
     AvSwitchCheckedIcon: {
       styleOverrides: {
-        root:{
+        root: {
           '& .Mui-checked': {
             '+ .MuiSwitch-track': {
               '&::before': {
                 left: 4,
-              }
-            }
-          }
-        }
-      }
+              },
+            },
+          },
+        },
+      },
     },
     MuiTab: {
       styleOverrides: {
@@ -2045,7 +2152,8 @@ export const lightTheme = {
         },
         selectLabel: {},
         selectIcon: {
-          fontSize: '1.5rem',
+          fontSize: tokens.iconSizeXxsmall,
+          marginRight: '8px'
         },
         spacer: {
           display: 'none',
@@ -2075,7 +2183,7 @@ export const lightTheme = {
         icon: {
           opacity: 1,
           transition: 'none',
-          fontSize: '14px',
+          fontSize: tokens.iconSizeXxsmall,
           color: tokens.colorTextSecondary,
           marginRight: '0px',
         },
@@ -2174,18 +2282,18 @@ export const lightTheme = {
       styleOverrides: {
         checkbox: {
           '&.Mui-checked': {
-            color: tokens.colorTextInverse
-          }
+            color: tokens.colorTextInverse,
+          },
         },
         content: {
           borderRadius: 0,
           padding: '8px 16px',
           alignItems: 'start',
           '&:hover': {
-            backgroundColor: tokens.colorActionHover
+            backgroundColor: tokens.colorActionHover,
           },
           '&:active': {
-            backgroundColor: `${tokens.colorSecondaryDark}26`
+            backgroundColor: `${tokens.colorSecondaryDark}26`,
           },
           '&.Mui-focused': {
             outlineWidth: tokens.borderWidthXs,
@@ -2206,18 +2314,18 @@ export const lightTheme = {
             },
             '&:active': {
               backgroundColor: tokens.colorPrimaryDarker,
-            }
+            },
           },
         },
         iconContainer: {
           color: tokens.colorSecondaryMain,
           paddingTop: '4px',
-          'svg': {
+          svg: {
             fontSize: 'inherit',
-            marginRight: '4px'
-          }
+            marginRight: '4px',
+          },
         },
-      }
+      },
     },
     MuiTypography: {
       defaultProps: {