+++ packages/avatar/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.3.4](https://github.com/Availity/element/compare/@availity/mui-avatar@0.3.3...@availity/mui-avatar@0.3.4) (2025-01-08)
+
+## [0.3.3](https://github.com/Availity/element/compare/@availity/mui-avatar@0.3.2...@availity/mui-avatar@0.3.3) (2025-01-06)
+
+### Dependency Updates
+
+* `mui-icon` updated to version `0.3.2`
 ## [0.3.2](https://github.com/Availity/element/compare/@availity/mui-avatar@0.3.1...@availity/mui-avatar@0.3.2) (2025-01-03)
 
 ### Dependency Updates