+++ packages/datepicker/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.4.56](https://github.com/Availity/element/compare/@availity/mui-datepicker@0.4.55...@availity/mui-datepicker@0.4.56) (2025-01-09)
+
+### Dependency Updates
+
+* `theme-provider` updated to version `0.4.55`
+## [0.4.55](https://github.com/Availity/element/compare/@availity/mui-datepicker@0.4.54...@availity/mui-datepicker@0.4.55) (2025-01-08)
+
+### Dependency Updates
+
+* `mui-textfield` updated to version `0.4.54`
+* `theme-provider` updated to version `0.4.54`
+## [0.4.54](https://github.com/Availity/element/compare/@availity/mui-datepicker@0.4.53...@availity/mui-datepicker@0.4.54) (2025-01-08)
+
+### Dependency Updates
+
+* `theme-provider` updated to version `0.4.53`
+## [0.4.53](https://github.com/Availity/element/compare/@availity/mui-datepicker@0.4.52...@availity/mui-datepicker@0.4.53) (2025-01-06)
+
+### Dependency Updates
+
+* `mui-icon` updated to version `0.4.52`
+* `mui-textfield` updated to version `0.4.52`
+* `theme-provider` updated to version `0.4.52`
 ## [0.4.52](https://github.com/Availity/element/compare/@availity/mui-datepicker@0.4.51...@availity/mui-datepicker@0.4.52) (2025-01-03)
 
 ### Dependency Updates