+++ packages/dialog/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.2.5](https://github.com/Availity/element/compare/@availity/mui-dialog@0.2.4...@availity/mui-dialog@0.2.5) (2025-01-08)
+
+### Dependency Updates
+
+* `mui-form-utils` updated to version `0.2.4`
+* `mui-textfield` updated to version `0.2.4`
+## [0.2.4](https://github.com/Availity/element/compare/@availity/mui-dialog@0.2.3...@availity/mui-dialog@0.2.4) (2025-01-06)
+
+### Dependency Updates
+
+* `mui-block-ui` updated to version `0.2.3`
+* `mui-form-utils` updated to version `0.2.3`
+* `mui-menu` updated to version `0.2.3`
+* `mui-alert` updated to version `0.2.3`
+* `mui-button` updated to version `0.2.3`
+* `mui-icon` updated to version `0.2.3`
+* `mui-textfield` updated to version `0.2.3`
 ## [0.2.3](https://github.com/Availity/element/compare/@availity/mui-dialog@0.2.2...@availity/mui-dialog@0.2.3) (2025-01-03)
 
 ### Dependency Updates