+++ packages/disclaimer/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.8](https://github.com/Availity/element/compare/@availity/mui-disclaimer@0.1.7...@availity/mui-disclaimer@0.1.8) (2025-01-06)
+
+### Dependency Updates
+
+* `mui-link` updated to version `0.1.7`
 ## [0.1.7](https://github.com/Availity/element/compare/@availity/mui-disclaimer@0.1.6...@availity/mui-disclaimer@0.1.7) (2025-01-03)
 
 ### Dependency Updates