+++ packages/form-utils/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.16.3](https://github.com/Availity/element/compare/@availity/mui-form-utils@0.16.2...@availity/mui-form-utils@0.16.3) (2025-01-08)
+
+### Dependency Updates
+
+* `mui-chip` updated to version `0.16.2`
+## [0.16.2](https://github.com/Availity/element/compare/@availity/mui-form-utils@0.16.1...@availity/mui-form-utils@0.16.2) (2025-01-06)
+
+### Dependency Updates
+
+* `mui-button` updated to version `0.16.1`
+* `mui-icon` updated to version `0.16.1`
+* `mui-checkbox` updated to version `0.16.1`
+* `mui-chip` updated to version `0.16.1`
 ## [0.16.1](https://github.com/Availity/element/compare/@availity/mui-form-utils@0.16.0...@availity/mui-form-utils@0.16.1) (2025-01-03)
 
 ### Dependency Updates