+++ packages/theme-provider/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.4.93](https://github.com/Availity/element/compare/@availity/theme-provider@0.4.92...@availity/theme-provider@0.4.93) (2025-01-09)
+
+### Dependency Updates
+
+* `theme` updated to version `0.4.92`
+## [0.4.92](https://github.com/Availity/element/compare/@availity/theme-provider@0.4.91...@availity/theme-provider@0.4.92) (2025-01-08)
+
+### Dependency Updates
+
+* `theme` updated to version `0.4.91`
+## [0.4.91](https://github.com/Availity/element/compare/@availity/theme-provider@0.4.90...@availity/theme-provider@0.4.91) (2025-01-08)
+
+### Dependency Updates
+
+* `theme` updated to version `0.4.90`
+## [0.4.90](https://github.com/Availity/element/compare/@availity/theme-provider@0.4.89...@availity/theme-provider@0.4.90) (2025-01-06)
+
+### Dependency Updates
+
+* `theme` updated to version `0.4.89`
 ## [0.4.89](https://github.com/Availity/element/compare/@availity/theme-provider@0.4.88...@availity/theme-provider@0.4.89) (2025-01-03)
 
 ### Dependency Updates