+++ packages/theme/CHANGELOG.md
@@ -2,6 +2,43 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.65.0](https://github.com/Availity/element/compare/@availity/theme@0.64.0...@availity/theme@0.65.0) (2025-01-09)
+
+
+### Features
+
+* **mui-stepper:** add scroll buttons ([5855cfb](https://github.com/Availity/element/commit/5855cfbdf8d43a709a37723c46b553631e115056))
+
+
+### Bug Fixes
+
+* **mui-stepper:** active state background color ([095cb98](https://github.com/Availity/element/commit/095cb98d72dc8d290a4128803919f87c82db611f))
+
+## [0.64.0](https://github.com/Availity/element/compare/@availity/theme@0.63.3...@availity/theme@0.64.0) (2025-01-08)
+
+### Dependency Updates
+
+* `design-tokens` updated to version `0.63.3`
+
+### Features
+
+* **theme:** use new iconsize tokens ([c1f9eec](https://github.com/Availity/element/commit/c1f9eecd1611bed86c07fa41965167fe278861aa))
+
+## [0.63.3](https://github.com/Availity/element/compare/@availity/theme@0.63.2...@availity/theme@0.63.3) (2025-01-08)
+
+
+### Bug Fixes
+
+* add hover border and fix accessibility ([7d8e7e4](https://github.com/Availity/element/commit/7d8e7e4136b266d42b3e453e57efe83f780a9e57))
+
+## [0.63.2](https://github.com/Availity/element/compare/@availity/theme@0.63.1...@availity/theme@0.63.2) (2025-01-06)
+
+
+### Bug Fixes
+
+* add hover color ([5001b77](https://github.com/Availity/element/commit/5001b77407b26de13d44b3ac3e39f05224aacc08))
+* various fixes for file selector ([26c651d](https://github.com/Availity/element/commit/26c651d7b8e89919fac335f233acbe528d10956a))
+
 ## [0.63.1](https://github.com/Availity/element/compare/@availity/theme@0.63.0...@availity/theme@0.63.1) (2025-01-03)